Extract `calc_imd_decile` to deciles.py
