+++ src/incognita/preprocessing/setup_reduce_onspd.py
@@ -5,6 +5,7 @@ from incognita.data.ons_pd_may_19 import ons_postcode_directory_may_19
 from incognita.logger import logger
 from incognita.logger import set_up_logger
 from incognita.utility import config
+from incognita.utility import deciles
 from incognita.utility import utility
 
 if __name__ == "__main__":
@@ -22,7 +23,7 @@ if __name__ == "__main__":
     logger.info("DEBUG - copied original data")
 
     # Add IMD Decile
-    data["imd_decile"] = utility.calc_imd_decile(data["imd"], data["ctry"], ons_postcode_directory_may_19).astype("UInt8")
+    data["imd_decile"] = deciles.calc_imd_decile(data["imd"], data["ctry"], ons_postcode_directory_may_19).astype("UInt8")
     logger.info("IMD Deciles added")
 
     # # TODO Needed?:
